==> notebooks/0--preliminaries+retrieval.ipynb <==
# Preliminaries

Here, we explain briefly how we organise our notebooks. Throughout all notebooks, we shall employ a custom module:

import ipynb_utils

It contains a configuration dictionary named `CFG` which holds global environment variables. This shall facilitate the process of storing and loading data between notebooks.

CFG = ipynb_utils.CFG

padding_length = max([len(k) for k in CFG])

print("CFG Dictionary:")
for k, v in CFG.items():
    print(f"  {k:{padding_length}} : {v}")

Finally, let us delineate the contents of our notebooks and the data upon which they respectively depend:

- `0--preliminaries+retrieval.ipynb`: This is the present notebook. The code that retrieves the data may be found in the subsequent section. The resulting file `df_processed.pkl` is requisite for all subsequent notebooks.
- `1--analysis.ipynb`: Performs basic exploratory data analysis and produces two processed data frames:
  - In the middle of the analysis, the file `df_duplicate.pkl` is produced, which is required for the excursory side analysis in `archive/1--analysis_hint-ignorant.ipynb`.
  - The file `df_processed.pkl`, produced at the end, is requisite for the notebook on modelling.
- `2--model.ipynb`: Contains the predictive Machine Learning models.

In the `archive/` directory, the reader may find two additional notebooks:

- `0--retrieval_kaggle.ipynb`: Describes the procedure to download the data via kaggle—with the required credentials.
- `1--analysis_hint-ignorant.ipynb`: Presents a data analysis in which we deliberately disregard the fact that the instructors intended to include a data duplication.

# Data Retrieval

In this notebook, we shall provide the code required to retrieve the data intended for analysis. It should be noted that appropriate credentials must be supplied to access the database.

We shall employ the following modules:

import os
from sqlalchemy import create_engine
import pandas as pd
from dotenv import load_dotenv

Subsequently, several useful variables for the retrieval process are introduced.

DATA_DIR = CFG["DATA_DIR"]

# Paths to which dataframes will be saved.
DF_PKL_PATH_TAR = os.path.join(DATA_DIR, "df_raw.pkl")
DF_CSV_PATH_TAR = os.path.join(DATA_DIR, "df_raw.csv") 

A brief examination of the given database, conducted within a GUI application such as DBeaver, reveals that the following SQL query yields the correct dataset:

QUERY = """
    SET SCHEMA 'diabetes';

    SELECT * FROM
            patient p
        LEFT JOIN
            blood_metrics bm
        ON p.id = bm.patientid
        LEFT JOIN
            pedigree_outcome po
        ON p.id = po.patientid
        LEFT JOIN
            skin s
        ON p.id = s.patientid
    ;
"""

We load the credentials required for database access from the `.env` file into Python.

load_dotenv()

DB_CONFIG = {
    "name": os.getenv("DB_NAME"),
    "scheme": os.getenv("DB_SCHEME"),
    "user": os.getenv("DB_USER"),
    "password": os.getenv("DB_PASSWORD"),
    "host": os.getenv("DB_HOST"),
    "port": os.getenv("DB_PORT")
}

DB_URI = (
    "{scheme}://{user}:{password}@{host}:{port}/{name}"
    .format(**DB_CONFIG)
)

We establish a connection with the database and load the data specified by the query into a pandas data frame.

db = create_engine(DB_URI)

with db.connect() as conn:
    df = pd.read_sql(QUERY, conn)

Let us confirm that the download process has been successful.

df.info()

df.sample(10)

The join operation in the SQL query may cause duplication of columns. Furthermore, the column `"patientid"` is identical to `"id"`. Such duplicates ought to be removed.

df = df.loc[:, ~df.columns.duplicated()]

if "patientid" in df.columns:
    df = df.drop(columns=["patientid"])

At last, the data frame is ready to be stored: To preserve the original data structure, we employ a pickle file; the csv version serves solely for direct visual inspection.

df.to_pickle(DF_PKL_PATH_TAR)
df.to_csv(DF_CSV_PATH_TAR, index=False)
